Added GUI email basic components
This was a simple addition, only added the graphical components to the
JDialogBox.  Functionality for the checkbox needs to be implemented
still (toggled false hides the input field, toggled true shows the email
input field). Additionally extracting the user input to apply it to the
event settings needs to be implemented.